+++ src-gui/src/renderer/components/pages/history/table/SwapMoneroRecoveryButton.tsx
@@ -107,7 +107,10 @@ export function SwapMoneroRecoveryButton({
   swap,
   ...props
 }: { swap: GetSwapInfoResponseExt } & ButtonProps) {
-  const isRecoverable = swap.state_name === BobStateName.BtcRedeemed;
+  const isRecoverable =
+    swap.state_name === BobStateName.BtcRedeemed ||
+    swap.state_name === BobStateName.XmrRedeemConstructed ||
+    swap.state_name === BobStateName.XmrRedeemPublished;
 
   if (!isRecoverable) {
     return <></>;